Moved SearchHits to shared library
